Moved verse-utils and verse-counts.
