+++ packages/web/src/features/translation/VerseSelector.tsx
@@ -1,6 +1,6 @@
 import { useState, KeyboardEvent } from 'react';
 import { useTranslation } from 'react-i18next';
-import { bookName, decrementVerseId, incrementVerseId, parseReference, parseVerseId } from '../../shared/verse-utils';
+import { bookName, decrementVerseId, incrementVerseId, parseReference, parseVerseId } from './verse-utils';
 import { Icon } from '../../shared/components/Icon';
 import TextInput from '../../shared/components/TextInput';
