+++ packages/web/src/features/translation/verse-counts.json
